add defensiveness, handle string values, rename files & refactor
