+++ src/components/Progress/ASABProgress.scss
@@ -1,3 +1,3 @@
-.progress-percentage {
+.asab-progress-percentage {
 	width: 2.4rem;
 }
